Connpass の処理が、"limit": 8589934588 という以上に大きな値が設定されていたことで落ちていた問題を修正
人数制限が 2147483647 x 4 という極端に大きな値が設定されていた。
https://beargguy3.connpass.com/event/289716/

+++ SlideCollector/src/main/java/tech/slideshare/connpass/Connpass.java
@@ -62,7 +62,7 @@ public record Connpass(
             @JsonProperty("hash_tag") String hashTag,
             @JsonProperty("started_at") ZonedDateTime startedAt,
             @JsonProperty("ended_at") ZonedDateTime endedAt,
-            @JsonProperty("limit") int limit,
+            @JsonProperty("limit") long limit,
             @JsonProperty("event_type") String eventType,
             @JsonProperty("series") Series series,
             @JsonProperty("address") String address,
